### 2019-01-11 #1163 @whedon
[PRE REVIEW]: pyscses - a python space charge site explicit solver
**Submitting author:** @georgiewellock (<a href="http://orcid.org/0000-0002-1068-5889">Georgina Wellock</a>)
**Repository:** <a href="https://github.com/bjmorgan/pyscses" target ="_blank">https://github.com/bjmorgan/pyscses</a>
**Version:** v 0.9.3.2
