### 2019-01-15 #1163 @labarba
@ncclementi — Would you be willing to volunteer a review for this submission to JOSS? You've experienced the review process as an author, so you'll know what is entailed, but have a look at the [documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) to imagine what it's like from the other side.
Let me know!
