### 2019-01-17 #1163 @labarba
:wave @vyasr — Have you reviewed for JOSS? You've published with us, I remember (quaternions!). Would you be able to help us out as second reviewer for this submission?
