### 2019-01-27 21a2355
Added missing DOIs
