+++ paper/paper.bib
@@ -104,6 +104,7 @@ doi = {10.1149/1.1507597}
         Number = {2},
         Pages = {185--187},
         Title = {Defect and transport properties of nanocrystalline {C}e{O}$_{2-x}$},
+        doi = {10.1063/1.117366},
         Volume = {69},
         Year = {1996}}
 
@@ -115,6 +116,7 @@ doi = {10.1149/1.1507597}
         Number = {3},
         Pages = {171--263},
         Title = {Ionic Conduction in Space Charge Regions},
+        doi = "10.1016/0079-6786(95)00004-E",
         Volume = {23},
         Year = {1995}}
 
