### 2019-01-27 #1209 @whedon
[REVIEW]: pyscses - a python space charge site explicit solver
**Submitting author:** @georgiewellock (<a href="http://orcid.org/0000-0002-1068-5889">Georgina Wellock</a>)
**Repository:** <a href="https://github.com/bjmorgan/pyscses" target ="_blank">https://github.com/bjmorgan/pyscses</a>
**Version:** v1.0.0
