### 2019-01-27 #1209 @labarba
👋 @georgiewellock  Please check your references list, aided by the bot's suggestions in the pre-review issue: https://github.com/openjournals/joss-reviews/issues/1163#issuecomment-457926514
