### 2019-01-28 7473aaa
Added JOSS status badge to README
