### 2019-01-28 #1209 @georgiewellock
> 👋 @georgiewellock Please check your references list, aided by the bot's suggestions in the pre-review issue: [#1163 (comment)](https://github.com/openjournals/joss-reviews/issues/1163#issuecomment-457926514)

I have updated the two references that have DOIs, but the other two do not. Frenkel (1946) does not appear to have been published digitally and the other is an unpublished paper we hope to have come out alongside the JOSS publication.
Latest commit: https://github.com/bjmorgan/pyscses/commit/21a2355cbf22d06e8e23829445d8cd525f19f11b
