### 2019-01-28 #1209 @labarba
In regards to the unpublished paper from the same authors, I want to encourage you to post the pre-print in a repository like [arXiv](https://arxiv.org), [ChemRxiv](https://chemrxiv.org) or similar. A citation to a non-public document is discouraged.
