### 2019-02-03 #1209 @vyasr
Nice work on the project! Here are some initial requests from my first pass of review, most of them shouldn't be too difficult.

- [ ] Remove the extraneous test_notebook folder in the root of the repo, as far as I can tell it's not being used any longer.
- [ ] Currently the unit test structure is a bit confusing. The Python files in the `tests/` directory fail because `pyscses` is still referred to as `project`, but even when I modify that naming issue I get other issues suggesting that those tests are using an old API. I'd remove those files, since they look like old versions of the files in `tests/unit_tests`.
- [ ] The tests in `tests/unit_tests` mostly work, except `test_grid.py` fails from trying to import `volume_from_grid`. It would be good to fix that, and then move these files directly into the `tests` directory since that's probably the more expected structure.
- [ ] The userguide looks pretty good overall, you should fill out or get rid of the README file there though. Also, you should look into [nbsphinx](https://nbsphinx.readthedocs.io/en/0.4.2/), which will allow you to post the userguide as part of your RTD documentation. People shouldn't have to clone the repository to access that guide, since it's probably the more useful part of the documentation than the API. Let me know if you need examples of how this is done.
- [ ] There's currently something wrong with the API documentation on ReadTheDocs. I'm not sure if it's an issue with RTD itself or with your build, but currently almost nothing actually shows up there. Please try and fix that, and let us know if there's some reason you can't get it working.
- [ ] The package appears to be built on a version of `sympy` that's many years old. The `mpmath` submodule of `sympy` was removed as of 0.7.6, and the current version is 1.3, so you'll need to add that as a separate dependency to make sure things still work.
- [ ] Make sure the test notebooks actually run successfully; I ran into a couple of (easy-to-fix) issues such as needing to install pandas (that should just be documented at the top of the notebook), having to make the directory for storing outputs, etc.
