### 2019-02-04 65f5e94
Removed unpublished paper citation
