### 2019-02-04 #1209 @bjmorgan
> In regards to the unpublished paper from the same authors, I want to encourage you to post the pre-print in a repository like [arXiv](https://arxiv.org), [ChemRxiv](https://chemrxiv.org) or similar. A citation to a non-public document is discouraged.

We have removed this unpublished reference (https://github.com/bjmorgan/pyscses/commit/65f5e940961175d1d029d41784eccb05ccedb67e). If it is ready for publishing as a preprint before this review is complete we will add it back at that stage.
