### 2019-02-04 bb4ab37
Updated unit tests
Starting to address review comments here:
https://github.com/openjournals/joss-reviews/issues/1209#issuecomment-460096293
