+++ README.md
@@ -38,14 +38,17 @@ pip install -e .
 ```
 ## Tests
 
-Jupyter notebooks that can be used to check the output of the calculations can be found in
-```
-pyscses/tests/test_notebooks
-```
-The test notebooks can be found on github [here](https://github.com/bjmorgan/pyscses/tree/master/tests/test_notebooks).
+The directory `tests/test_notebooks` contains a set of Jupyter notebooks with specified outputs, that can be run to test code functionality. The test notebooks can be found on GitHub [here](https://github.com/bjmorgan/pyscses/tree/master/tests/test_notebooks).
 
 There are four directories with varying conditions. In each there is a Jupyter notebook which can be run. The input for the calculations is stored in the `input_data` directory and the output for the calculations will be stored in the `generated_data` directory and can be compared to the verified data in the `expected_outputs` directory. A list of the input parameters used in the notebooks is reiterated in each of the four test folders in the `input_parameters` file. 
 
+### Unit tests
+
+Limited unit tests are contained in the top `tests` directory. These can be run using
+```
+cd tests
+python -m unittest discover
+```
 ## Documentation
 Once installed, the `pyscses` code is imported into, and run using a [Jupyter notebook](http://jupyter-notebook.readthedocs.io/en/latest/#).
 An overview of the capabilities of `pyscses`, along with example code for running the code and varying the simulation conditions can be found in
