+++ .travis.yml
@@ -1,0 +1,23 @@
+branches:
+  only:
+    - master
+    - development
+language: python
+notifications:
+  email: false
+python:
+  - "3.5"
+  - "3.5-dev" # 3.5 development branch
+  - "3.6"
+  - "3.6-dev" # 3.6 development branch
+matrix:
+  include:
+    - python: 3.7
+      dist: xenial
+      sudo: true
+# command to install dependencies
+install: pip install -r requirements.txt
+# command to run tests
+script: coverage run --rcfile=.coveragerc -m unittest discover
+after_success:
+  - codeclimate-test-reporter --token $CODECLIMATE_REPO_TOKEN
