### 2019-02-04 bbe2712
Added coverage dependencies to travis config
