### 2019-02-04 6fdb52a
Added Travis-CI link to README
