+++ docs/requirements.txt
@@ -1,2 +1,4 @@
 Sphinx>=1.6.3
+sphinx_rtd_theme
+nbsphinx
 numpy
