### 2019-02-04 1abbde0
Updated docs requirements
