+++ docs/requirements.txt
@@ -1,3 +1,4 @@
+-r ../requirements.txt
 Sphinx>=1.6.3
 sphinx_rtd_theme
 nbsphinx
