### 2019-02-04 #1209 @bjmorgan
API documentation appears to be fixed after the latest commit (https://github.com/bjmorgan/pyscses/commit/1abbde059886511f82d6cba1056916a31e97e2a7).

e.g. https://pyscses.readthedocs.io/en/latest/pyscses.html
