### 2019-02-05 bjmorgan/pyscses#7 @ncclementi
Wording and problems in Theory.ipynb notebook
1. I was reading the notebook and I landed on the paragraph below that is hard to understand since the word "boundary" appears four times, is there a way to rephrase this?

"For solid electrolytes, the Poisson-Boltzmann equation is commonly solved by treating the grain **boundary** via the **boundary** condition by setting the grain **boundary** potential. However the grain **boundary** can be treated explicitly..."

2. For the record, every time you use "However"  at the beginning of a sentence you need to put a comma afterwards. (I caught some and submit a PR same with typos, but I might have missed some). For the misspelled words, I's suggest you install a spell checker for jupyter notebooks, there are many options out there.

3. The image `<img src="Figures/irregulargrid.png">`  doesn't exists and therefore doesn't render.
