### 2019-02-05 bjmorgan/pyscses#10 @ncclementi
Time for running example notebooks?
Hello there,

How long are the Examples notebooks on the userguide are suppose to take? I'd like to have a reference so I can run them. I've been trying to run Example 1 and the times has gone over 10 min is that expected?

Thanks
