### 2019-02-05 #1209 @ncclementi
Nice job people, here are my comments, requests and questions so far on each part that I'm suppose to review. I haven't been able to get to the tests part yet, but I saw @vyasr made some comments already.

I divided my requests, questions and comments by following the structure that is given as for the review, you can find it on the first comment.

### General comments, requests and questions regarding:

**Version**: Does the release version given match the GitHub release (v 0.9.3.2)?

It matches now but if you modify things after the review, you will want to do a release and in that it case it won't match, how do we deal with this? @labarba are there any suggestion regarding this?

**Authorship**: Has the submitting author (@georgiewellock) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

According to the joss reviewer guidelines, I need to check weather the submitting author has made "substantial contribution" to the submitted software (as determined by the commit history). https://joss.readthedocs.io/en/latest/review_criteria.html

The submission was made by the username **@georgiewellock** but this account has no
commits in the repository contributors section https://github.com/bjmorgan/pyscses/graphs/contributors

Can you explain how you meet the criteria?

### Functionality

**Installation**: Does installation proceed as outlined in the documentation?

I tried to installed via `pip` and had no problem until I tried to run and had problem with `mpmath` dependency.

Setting up the notebook to run calculations:

- `ImportError: cannot import name 'mpmath'`
@vysar already point this out. You need to add this a separate dependency to make sure things work. I've saw this modified as `import mpmath` but after installing `mpath` separately and pulling the latests changes from the repo I tried to install `pyscses` following the instructions but I still can't run the setup notebook. This is because all
the installation instructions rely on `pip` and `pip` doesn't have the latest version.

You should add an option of installing it directly using the `setup.py` and that way it works.

Also pandas seems to be needed to run the notebooks too, but it's not listed as a dependency, it should be added.

Lastly, the software is run by using Jupyter notebooks, even though this is mentioned later in the documentation, might be useful to mention it at the installation level.

**Functionality**: Have the functional claims of the software been confirmed?

To run the examples and tests I need certain notebooks that are in the repo that I
can't download all together unless I clone the repo.

Is there a better way to do this? you should put a note or description on how to get the examples and tests. Since we can't run the notebooks from github directly it would be useful to add a comment on where in the repo are located instead of putting "definitions and example code can be found in the `userguide`." where userguide is a specific link.

Make sure that all your notebooks run clean and it would be useful to have information on the time it takes to run some of the examples.

- I open an issue on the repo so you can give me an estimate on the timing and maybe on the future add it to the documentation, since I've been running Example 1 for about 15 min and I don't know if this is the expected behavior.

- Running notebook:
Cell number 6:
```---------------------------------------------------------------------------
NameError                                 Traceback (most recent call last)
<ipython-input-6-67db249c57f8> in <module>()
      4     for t in temp:
      5
----> 6         defect_species = { l : DefectSpecies( l, v, m, mob ) for l, v, m, mob in zip( defect_labels, valence, m, mobilities ) }
      7
      8         all_sites = Set_of_Sites.set_of_sites_from_input_data( data, [grid_x_min, grid_x_max], defect_species, site_charges, core_models, t )

NameError: name 'mobilities' is not defined
```

**Performance**: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

There are no claims of performance so I checked the item off.

### Documentation

**A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I think that the summary on the readme doesn't clearly say what problems the software is
designed to solve. There is a sentence that opens the paragraph:

"pyscses is a Python module that implements a site-explicit, one-dimensional Poisson-Boltzmann solver, used for modelling ionic space charge properties in solid materials."

Is think you should add more that explains the target problems to general audience and why this problems are important.

**Installation instructions**: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Read comments on Installation made in the Functionality section.

**Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There are examples on how to use it. Haven't been able to run them yet. I need
a better idea on the time they take to run (explained in Functionality section above)

**Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

There is API documentation in the webpage for each class, and the documentation of each me
method.

**Automated tests**: Are there automated tests or manual steps described so that the function of the software can be verified?

Haven't get to this part yet, but I see @vysar made some comments on it so I will work on this once his suggestions are included.

**Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

There are no guidelines on how to contribute to the software, report issues or problems with the software or how to seek support.

### Software paper

Comments following the checklist:

Keeping in mind that according to the authors guidelines "A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience"

**A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I'm not sure if I understand where is the statement of need presented in the software and who is the target audience.

At the beginning of the the paper you mention what `pyscses` is used for:
"modelling ionic space-charge regions in crystalline solids."

Can you break down this for a more general audience?

The following two paragraphs intend to explain more about the problem but it is hard to understand, at least for someone that is not from the area. I get a bit lost due to the specific vocabulary and some of the sentences are too long. For example, the sentence that starts with "For example, ..." in the second paragraph goes for five lines.

and at the end of the second paragraph you say:

 "In the case of solid electrolytes, understanding space-charge formation at grain boundaries and interfaces is a key challenge in developing a theoretical description of the role of crystalline microstructure on macroscopic ion transport (ionic conductivities) [@KimEtAl_PhysChemChemPhys2003]."

This tells me why "understanding space-charge formation..." is important but not what is the purpose of the software. I'm missing the connection between the problem, its importance and the software itself.

**References**:

Regarding the references, you have an "in preparation" paper but it is cited in the numerical model. Are you using `pyscses` for this research, it is not clear to me? If so, it is important to mention it. According to authors guidelines "Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it".

#### General comments regarding the paper:

If we go back to the documentation on [Submitting a paper to JOSS](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)

I think the paper is too long. You might want to review some of the examples of published papers, or even though the one in the documentation.

There are multiple parts where you explain some of the software functionality and workflow (End of numerical model section and "Typical workflow section"), and this should not be included. According to the review criteria:

"The paper should not include software documentation such as API functionality,
as this should be outlined in the software documentation."
https://joss.readthedocs.io/en/latest/review_criteria.html#review-items

*Numerical model section*

Regarding this section, I don't think is necessary to have it on the paper since most of it is explained with the same words in the `Theory.ipynb` notebook and example notebooks.

*Calculated properties, approximations and limitations sections*

I think these shouldn't be separate sections. I would suggest to rephrase all the information in this sections, add it to the summary and remove the math.

**Remember** that the core of the paper should, according to the guidelines, contain :

- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software

Acknowledgments:

There is a grant code missing: (EPSRC grant code?).
