### 2019-02-05 #1209 @labarba
@ncclementi — You say that you opened an issue on the submission repo. Can you add a link to that issue? The cross-links are nice to have on the Review thread. Thanks!

Regarding the software version, after changes stemming from the review, the authors will have to up the version, and we'll use the superpowers of our bot to update it (`@whedon set v1.x as version`).
