### 2019-02-05 #1209 @ncclementi
Yes, sure! Here are the links of the issues I opened:
https://github.com/bjmorgan/pyscses/issues/7
https://github.com/bjmorgan/pyscses/issues/10
