### 2019-02-06 #1209 @bjmorgan
@ncclementi @vyasr Thank you both for your very thorough comments. We will begin working through addressing these points.
