### 2019-02-06 #1209 @bjmorgan
@vyasr
> Remove the extraneous test_notebook folder in the root of the repo, as far as I can tell it's not being used any longer.
Done in commit https://github.com/bjmorgan/pyscses/commit/969d7be62bd4fa8777383a274de2c6fa72d251ea
