### 2019-02-06 #1209 @bjmorgan
@vyasr
> * [x] Currently the unit test structure is a bit confusing. The Python files in the `tests/` directory fail because `pyscses` is still referred to as `project`, but even when I modify that naming issue I get other issues suggesting that those tests are using an old API. I'd remove those files, since they look like old versions of the files in `tests/unit_tests`.
> * [x]  The tests in `tests/unit_tests` mostly work, except `test_grid.py` fails from trying to import `volume_from_grid`. It would be good to fix that, and then move these files directly into the `tests` directory since that's probably the more expected structure.

Both addressed in commit https://github.com/bjmorgan/pyscses/commit/969d7be62bd4fa8777383a274de2c6fa72d251ea

Although we only have very limited unit tests at the moment, we have also added continuous integration tests using Travis CI, to avoid these tests getting out of sync with the project API in the future.
