### 2019-02-06 d3d61f9
Version number bump to 0.9.3.4
