+++ pyscses/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.9.3.2'
+__version__ = '0.9.3.3'
