### 2019-02-06 #1209 @bjmorgan
@vyasr
> * The package appears to be built on a version of `sympy` that's many years old. The `mpmath` submodule of `sympy` was removed as of 0.7.6, and the current version is 1.3, so you'll need to add that as a separate dependency to make sure things still work.

`requirements.txt` has been updated.

@ncclementi
> I tried to installed via `pip` and had no problem until I tried to run and had problem with `mpmath` dependency.
>
> Setting up the notebook to run calculations:
>
> * `ImportError: cannot import name 'mpmath'`
  @vysar already point this out. You need to add this a separate dependency to make sure things work. I've saw this modified as `import mpmath` but after installing `mpath` separately and pulling the latests changes from the repo I tried to install `pyscses` following the instructions but I still can't run the setup notebook. This is because all
  the installation instructions rely on `pip` and `pip` doesn't have the latest version.
>
> You should add an option of installing it directly using the `setup.py` and that way it works.

See above re: updates to `requirements.txt`.

We have also updated the version on PyPI to let you test installation via `pip`.

The installation instructions in the top level `README` contain instructions for alternate installation directly from the GitHub source using `setup.py`:

> ## Installation
>
> The simplest way to install pyscses is to use pip to install from PyPI
> ```
> pip install pyscses
>```
> Alternatively, you can download the latest release from GitHub, and install directly:
>
> ```
> cd pyscses
> pip install -e .
> ```
> which installs an editable (-e) version of pyscses in your userspace.
>
> Or clone the latest version from GitHub with
>
> ```
> git clone git@github.com:bjmorgan/pyscses.git
> ```
> and install the same way.
>
> ```
> cd pyscses
> pip install -e .
> ```
