### 2019-02-06 #1209 @bjmorgan
@ncclementi
> **Authorship**: Has the submitting author (@georgiewellock) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> According to the joss reviewer guidelines, I need to check weather the submitting author has made "substantial contribution" to the submitted software (as determined by the commit history). https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission was made by the username **@georgiewellock** but this account has no
commits in the repository contributors section https://github.com/bjmorgan/pyscses/graphs/contributors
> Can you explain how you meet the criteria?

If you look at the full commit history at https://github.com/bjmorgan/pyscses/commits/master you can see the majority of commits are from `glw33` who is @georgiewellock. It seems @georgiewellock's local git setup is using a different username than their GitHub account.

@labarba I'd like to flag a possible point of confusion in your reviewer / author instructions: Your guide for reviewers at present starts by saying that the commit history should be used to verify significant contributions, but then later says "**active project direction and other forms of non-code contributions are [significant]**". I can see that if someone has a substantial code contribution they should be considered as an author, but it is not necessarily clear at the moment that a significant commit history is not the only requirement for authorship.
