### 2019-02-06 8e36373
Added guidelines for contributing to README.md
