### 2019-02-06 #1209 @bjmorgan
@ncclementi

> **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> There are no guidelines on how to contribute to the software, report issues or problems with the software or how to seek support.

We have added guidelines for contributing / asking for support / reporting issues to the top level `README`: https://github.com/bjmorgan/pyscses/commit/8e36373c28ef79918970dd4b62890dec1ebdc2fb
