### 2019-02-06 #1209 @ncclementi
@bjmorgan

Regarding installation: I will have to give it another try, because when I tried to installed (after the requirements were updated) I wasn't able to get it working doing:
Or clone the latest version from GitHub with

```
git clone git@github.com:bjmorgan/pyscses.git
```
and install the same way.

```
cd pyscses
pip install -e .
```
The way it worked for me was by doing:
`python setup.py install`

Regarding your request: "Can you please add further comments and suggestions / those we have not yet addressed as issues on the main repository, so we can track which we have / have not addressed?"

I'm not sure if I understand what you mean. I think went over all what is required to review on the big comment I made above, did I miss something?. I know that my path on the tests is pending but I will take a look at it soon. Are you referring to that, or something different?
