### 2019-02-06 #1209 @bjmorgan
@ncclementi
Can you please post your errors when you try to install using `pip install -e .`?

Regarding opening specific issues on the main repository: you have both provided comprehensive feedback on your first pass, and I want to make sure we don't miss any of the issues you have raised. It will probably be easier to track individual discussion threads in separate issues on the main repo than having things all mixed together here.
