### 2019-02-06 bjmorgan/pyscses#11 @bjmorgan
Suggested edits to JOSS paper
Comments from @ncclementi here --> https://github.com/openjournals/joss-reviews/issues/1209:

### Software paper
Comments following the checklist:

Keeping in mind that according to the authors guidelines "A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience"

**A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I'm not sure if I understand where is the statement of need presented in the software and who is the target audience.

At the beginning of the the paper you mention what `pyscses` is used for:
"modelling ionic space-charge regions in crystalline solids."

Can you break down this for a more general audience?

The following two paragraphs intend to explain more about the problem but it is hard to understand, at least for someone that is not from the area. I get a bit lost due to the specific vocabulary and some of the sentences are too long. For example, the sentence that starts with "For example, ..." in the second paragraph goes for five lines.

and at the end of the second paragraph you say:

"In the case of solid electrolytes, understanding space-charge formation at grain boundaries and interfaces is a key challenge in developing a theoretical description of the role of crystalline microstructure on macroscopic ion transport (ionic conductivities) [@KimEtAl_PhysChemChemPhys2003]."

This tells me why "understanding space-charge formation..." is important but not what is the purpose of the software. I'm missing the connection between the problem, its importance and the software itself.

**References**:

Regarding the references, you have an "in preparation" paper but it is cited in the numerical model. Are you using `pyscses` for this research, it is not clear to me? If so, it is important to mention it. According to authors guidelines "Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it".

#### General comments regarding the paper:
If we go back to the documentation on [Submitting a paper to JOSS](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)

I think the paper is too long. You might want to review some of the examples of published papers, or even though the one in the documentation.

There are multiple parts where you explain some of the software functionality and workflow (End of numerical model section and "Typical workflow section"), and this should not be included. According to the review criteria:

"The paper should not include software documentation such as API functionality,
as this should be outlined in the software documentation."
https://joss.readthedocs.io/en/latest/review_criteria.html#review-items

_Numerical model section_

Regarding this section, I don't think is necessary to have it on the paper since most of it is explained with the same words in the `Theory.ipynb` notebook and example notebooks.

_Calculated properties, approximations and limitations sections_

I think these shouldn't be separate sections. I would suggest to rephrase all the information in this sections, add it to the summary and remove the math.

**Remember** that the core of the paper should, according to the guidelines, contain :

* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
* A clear statement of need that illustrates the purpose of the software

Acknowledgments:

There is a grant code missing: (EPSRC grant code?).
