### 2019-02-06 #1209 @bjmorgan
@vyasr
>  The userguide looks pretty good overall, you should fill out or get rid of the README file there though. Also, you should look into nbsphinx, which will allow you to post the userguide as part of your RTD documentation. People shouldn't have to clone the repository to access that guide, since it's probably the more useful part of the documentation than the API. Let me know if you need examples of how this is done.

From my initial playing around, if we want to add these `*.ipynb` notebooks to the Sphinx RTD documentation, they need to be moved into the `docs/source` directory. I feel this makes them difficult to find if people want to access and run them as Jupyter notebooks. What are your thoughts on providing web-viewable versions using e.g. nbviewer --> https://nbviewer.jupyter.org/github/bjmorgan/pyscses/tree/master/userguides/notebooks/
