### 2019-02-06 bjmorgan/pyscses#10 @bjmorgan
@georgiewellock Can you add approximate timings to the userguides README, and note in the notebooks where there are any calculations that take a long time to run?
