### 2019-02-06 #1209 @labarba
> you have both provided comprehensive feedback on your first pass, and I want to make sure we don't miss any of the issues you have raised. It will probably be easier to track individual discussion threads in separate issues on the main repo than having things all mixed together here.

@bjmorgan — Indeed, the reviewers have provided a detailed report of the issues you may need to address. For a few concrete ones, they opened a new issue on you software repository. With the rest, as it is posted here, I will request that you take the time to create a checklist of all the items that need addressing, and you go about checking them off (and open issues in your own repo as needed). I don't support requesting that the reviewers do this job for you.
