### 2019-02-06 #1209 @bjmorgan
> @bjmorgan — Indeed, the reviewers have provided a detailed report of the issues you may need to address. For a few concrete ones, they opened a new issue on you software repository. With the rest, as it is posted here, I will request that you take the time to create a checklist of all the items that need addressing, and you go about checking them off (and open issues in your own repo as needed). I don't support requesting that the reviewers do this job for you.

@labarba We're happy to copy the more complex points to issues on the main repo to help track them.
