### 2019-02-07 9295b70
Tidying docstrings
