+++ docs/build/html/_sources/index.rst.txt
@@ -1,22 +1,0 @@
-.. pyscses documentation master file, created by
-   sphinx-quickstart on Fri Mar  9 11:37:43 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-pyscses - PYthon Space-Charge Site-Explicit Solver
-====================================================
-
-.. include:: readme.rst
-
-.. toctree::
-   :caption: API documentation
-   :maxdepth: 3
-   
-   modules
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`

[…]
