### 2019-02-07 9829b8c
corrected code contributions
