+++ README.md
@@ -96,7 +96,7 @@ If you think you have found a bug, please report it on the Issue Tracker. This i
 
 We welcome your help in improving and extending the package with your own contributions. This is managed through GitHub pull requests; for external contributions we prefer the "fork and pull" workflow, while core developers use branches in the main repository:
 
-- First open an Issue to discuss the proposed contribution. This discussion might include how the changes fit Sumo's scope and a general technical approach.
+- First open an Issue to discuss the proposed contribution. This discussion might include how the changes fit pyscses' scope and a general technical approach.
 - Make your own project fork and implement the changes there. Please keep your code style compliant with PEP8.
 - Open a pull request to merge the changes into the main project. A more detailed discussion can take place there before the changes are accepted.
 
