### 2019-02-07 b845a5d
fixed running.ipynb and added expected time for completion
