### 2019-02-07 8ceeab3
Approximate running times and iMac specs added to README and notebooks
