### 2019-02-07 bjmorgan/pyscses#10 @georgiewellock
I have added the approximate timings and machine specifications to both the README and the userguide notebooks: https://github.com/bjmorgan/pyscses/commit/8ceeab33d066e1a50ab6180a362cac95fac6f1db

Example notebooks 1 and 2 take significantly longer to run due to the Gouy-Chapman conditions. These require the Poisson-Boltzmann solver to be run repeatedly in order to minimise the difference between the mole fractions we pass in and the mole fractions which are calculated. Example 3 onwards (should) only take a few minutes.
