### 2019-02-07 2c2965e
fixed irregulargrid.png and reformatted some text. Ran spellchecker and corrected errors
