### 2019-02-07 bjmorgan/pyscses#7 @georgiewellock
I have reworded the paragraph to avoid using 'boundary' so frequently. I have imported used a spell checker and corrected the mistakes flagged up and I have fixed the irregular grid figure so that it renders in the notebook. These changes can be found in commit https://github.com/bjmorgan/pyscses/commit/2c2965e360f5479e5a8e0b9eb4d717482916e8bf.
