+++ requirements.txt
@@ -1,3 +1,4 @@
+pandas
 numpy
 mpmath
 scipy
