### 2019-02-07 19665a7
updated userguide README with dependencies
