+++ userguides/requirements.txt
@@ -1,0 +1,4 @@
+pandas
+numpy
+mpmath
+scipy
