### 2019-02-07 eeea938
updated userguide README with links to notebooks
