### 2019-02-07 bf47314
updated userguide README with links to notebooks
