### 2019-02-07 9cb1d1a
updated userguide README with links to notebooks
