### 2019-02-07 89210c1
updated userguide README with links to notebooks
