### 2019-02-07 232b8a4
removed form_continuum_sites_new as unused function
