### 2019-02-07 fa41182
updated userguide link
