### 2019-02-07 2d4215c
updated userguide description
