### 2019-02-07 055bbf3
updated docstrings for attributes rather than returns
