+++ pyscses/calculation.py
@@ -60,7 +60,7 @@ class Calculation:
 
         Args:
             input_mole_fractions (list): Mole fractions for each of the species used in the iterative Poisson-Boltzmann solver.
-            approximation (?): TODO
+            approximation (str): The defect mobility approximation. Either 'mott-schottky' to enforce only a single mobile defect, or 'gouy-chapman' to allow all defect species to redistribute. 
    
         Returns:
             list: Mole fractions that are calculated from the iterative Poisson-Boltzmann solver.
@@ -91,15 +91,12 @@ class Calculation:
 
         
     def mole_fraction_correction( self, target_mole_fractions, approximation, initial_guess ):
-        """Starting from an initial guess for the appropriate input mole fractions, minimises the error between the target bulk mole fraction and the output mole fraction from the iterative Poisson-Boltzmann solver. 
+        """Starting from an initial guess for the appropriate input mole fractions, minimises the error between the target bulk mole fraction and the output mole fraction from the iterative Poisson-Boltzmann solver. The output is stored as a Calculation attribute. Calculation.initial_guess (list): The optimum values to be used as the input mole fractions for the iterative Poisson-Boltzmann solver so that the output bulk mole fractions are the target bulk mole fractions. 
 
         Args:
             target_mole_fractions (list): The value that the mole fractions should be in the bulk. 
-            approximation (?): TODO
-            initial_guess (?): TODO
-
-        Returns:
-            list: The optimum values to be used as the input mole fractions for the iterative Poisson-Boltzmann solver so that the output bulk mole fractions are the target bulk mole fractions. 
+            approximation (str): The defect mobility approximation. Either 'mott-schottky' to enforce only a single mobile defect, or 'gouy-chapman' to allow all defect species to redistribute. 
+            initial_guess (list): Values for an initial guess for the defect mole fractions used in the error minimisation. 
 
         """
         self.initial_guess = initial_guess
@@ -185,16 +182,16 @@ class Calculation:
 
     def solve( self, approximation ):
         """
-        Self-consistent solving of the Poisson-Boltzmann equation. Iterates until the convergence is less than the convergence limit.
+        Self-consistent solving of the Poisson-Boltzmann equation. Iterates until the convergence is less than the convergence limit. The outputs are stored as Calculation attributes.
+        Calculation.phi (array): Electrostatic potential on a one-dimensional grid. 
+        Calculation.rho (array): Charge density on a one-dimensional grid.
+        Calculation.niter (int): Number of iterations performed to reach convergence.
+
 
         Args:
             approximation (str): Approximation used for the defect behaviour.
                                  'mott-schottky' - Some defects immobile / fixed to bulk mole fractions.
                                  'gouy-chapman' - All defects mobile / able to redistribute.
-        Returns:
-            numpy.array: Electrostatic potential on a one-dimensional grid. 
-            float: Charge density on a one-dimensional grid.
-            int: Number of iterations performed to reach convergence.
 
         """
         poisson_solver = MatrixSolver( self.grid, self.dielectric, self.temp, boundary_conditions=self.boundary_conditions )
@@ -227,13 +224,10 @@ class Calculation:
         self.niter = niter
 
     def form_subgrids( self, site_labels ):
-        """Creates a `pysces.Grid` object for each species in the system.
+        """Creates a `pysces.Grid` object for each species in the system. The output is a dictionary of separate Grid classes for the different site species and is stored as Calculation.subgrids. 
 
         Args:
[…]
