### 2019-02-08 bjmorgan/pyscses#10 @ncclementi
Great, thanks for adding that I will be running the examples today!
