### 2019-02-08 bjmorgan/pyscses#13 @ncclementi
`userguides/notebooks/.ipynb_checkpoints`
Even though you have the `userguides/notebooks/.ipynb_checkpoints` in the `.gitignore` you still have some `ipynb_checkpoints` in the repository. You might want to do a `git rm` to remove them.
