### 2019-02-08 #1209 @ncclementi
Issue opened regarding presence of `.ipynb_checkpoints`
https://github.com/bjmorgan/pyscses/issues/13
