### 2019-02-08 bjmorgan/pyscses#7 @georgiewellock
Yes, sorry I forgot to push up the new image. I have done so now: https://github.com/bjmorgan/pyscses/commit/7dba222392bfcbc750fc6cba6a13789ab29bc18e
