### 2019-02-08 07e17a4
removed .ipynb_checkpoints
