### 2019-02-08 bjmorgan/pyscses#13 @georgiewellock
Thanks, removed here: https://github.com/bjmorgan/pyscses/commit/07e17a4e8fb5d72d9f8bc6b189f714f9588b66bf
