### 2019-02-08 bjmorgan/pyscses#7 @ncclementi
Thanks, I think you cover everything on this issue so feel free to close it.
Great job!
