### 2019-02-08 bd0572e
Updated .gitignore
