### 2019-02-08 bjmorgan/pyscses#13 @bjmorgan
Additional pruning of `**/.ipynb_checkpoints`: https://github.com/bjmorgan/pyscses/commit/bd0572e23e7e25389a00e8ceb0a1cb6c884d1f63
