### 2019-03-05 #1209 @ncclementi
In my case, I saw you addressed my comments on the software but it is still pending my comments on the paper itself. I will give another path to the rest of the points but I'd like the authors to address my comments on the `paper.md`.
