### 2019-03-05 #1209 @vyasr
@bjmorgan regarding the notebooks, I see your point about wanting to keep them easily visible rather than hiding them in a docs folder. One solution would be to move the notebooks to `docs/source/examples` or so, and then symlink them to your tests directory (run `ln -s docs/source/examples userguides` in the root of the repository). Would that address your concern? I think that most users would be greatly aided by actually seeing these examples on RTD, since these days that's where most Python code hosts explanatory documentation in addition to API docs.

Having said that, I do think nbviewer is also a good option if you dislike my suggestion. Another alternative you may want to consider is mybinder. The service isn't perfect and does go down, so I would definitely only do that in addition to RTD/nbviewer, but it's really nice to be able to execute the notebooks directly online. That's just a suggestion though, not a hard request for this paper.
