### 2019-03-05 #1209 @vyasr
The new contributing guidelines look good. I would suggest converting the "Issue Tracker" and "pull request" text into links.
