### 2019-03-05 #1209 @vyasr
Regarding the contents of the paper, I'm mostly in agreement with @ncclementi. The paper is definitely too long at present, so I'll try to give pointers on where and how to cut as I make other comments.

I found the summary to be comprehensible, but I agree that it's probably a little too technical for a general, non-specialist audience unfamiliar with the behavior of defects and grain boundaries in crystal structures. I would probably summarize the first two paragraphs with something along the lines of:
"Most real crystalline materials are characterized by defects, imperfections that violate the perfect symmetries that define the crystal. For example, in many crystals, two locally ordered regions that are oriented differently meet at what is called a grain boundary, where the two crystalline pieces fail to fit perfectly together. In ionic crystals, thermodynamic arguments show that charged point defects will migrate towards interfaces and grain boundaries, leading to the formation of surface space-charges, or regions of net charge. This segregation can impact various material properties, e.g. (INSERT JUST ONE EXAMPLE, SKIP THE REST). Therefore, understanding the distributions of point-charge defects is critical to understanding the properties of a material. (INSERT SENTENCE: WHY DEVELOP pyscses? ARE THERE NO OTHER GOOD TOOLS, NOTHING IN PYTHON, ETC)."

I don't quite agree with @ncclementi that you should remove this whole section, I think it's nice to have some of it. You can definitely combine most of the numerical model and calculated properties sections, and remove most of the math. I would probably drop the first two paragraphs and start the section with the paragraph on "Finding the equilibrium distribution". Keep the first equation there, then drop the rest and just assert that this leads to a system involving a modified PB equation that can be solved in a self-consistent fashion. You can point to your docs and userguides for more information. I actually like the bulleted list at the end of this section, and I would keep it since I don't consider that API level information. I'd remove the first two paragraphs of the typical workflow section, then just have one sentence for each of the calculable properties without section headers or math.

I'm curious to hear @ncclementi's thoughts on whether she thinks that's still too much information.
