### 2019-03-05 #1209 @ncclementi
@vyasr @bjmorgan Regarding the notebooks and mybinder, I'm not sure if that would work, I run the examples on my machine `Intel(R) Core(TM) i7-3820 CPU @ 3.60GHz` and they took a really long time. Some of them over an hour of computation or even more.
A good option would be selecting smaller problems and see how long they take on my binder. But I agree that being able to interact with the example notebooks is a good idea.
