### 2019-03-05 #1209 @ncclementi
@vyasr @bjmorgan I think you can leave some of the numerical method section but maybe a summary and something not that technical. The reason why I suggested to remove it it is because this information is already present, almost verbatim  in the Theory.ipynb notebook and example notebooks. Maybe you can point to that document instead.

Remember that the core of the paper should, according to the guidelines, contain :
-A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
-A clear statement of need that illustrates the purpose of the software

Then, any other thing you add I'd try to keep it short and for a general audience. You can go more into detail in the documentation and notebooks in the repo.
