### 2019-03-05 #1209 @vyasr
Yes, I definitely think that MyBinder is not a good idea for very complex examples. I have not yet tried running the examples again since the first round of review. I'll do that once some more of these changes have been addressed so that you don't have too many comments at once, but I concur that putting some smaller problems for MyBinder would be nice if they can solve fast.

I'd say let's see what the proposed changes look like, then we can have a discussion about cutting further if needed. I think having some technical details is helpful to orient readers with some background, perhaps in related fields where they won't make the connection immediately without seeing a little detail. It's good for appealing to potential users who are close to the field without really being subject-matter experts.
