### 2019-03-08 afd6efa
Added nbviewer links to userguide README
