+++ userguides/README.md
@@ -10,16 +10,16 @@ pip install -r requirements.txt
 
 ## Notebooks
 
-Each notebook can be viewed using [nbviewer](https://nbviewer.jupyter.org) by following each <span style="color:orange">nbviewer</span> link.
-- [Theory](notebooks/Theory.ipynb) ([<span style="color:blue">nbviewer</span>](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Theory.ipynb))
-- [Setting up the notebook](notebooks/Setting_up.ipynb) 
-- [Running the calculation](notebooks/Running.ipynb)
-- [Example 1 - continuum vs. site explicit and boundary conditions](notebooks/Ex_1_BC.ipynb)
-- [Example 2 - Mott-Schottky vs. Gouy-Chapman conditions](notebooks/Ex_2_MSGC.ipynb)
-- [Example 3 - Calculating the grain boundary resistivity and activation energy](notebooks/Ex_3_Res.ipynb)
-- [Example 4 - Comparison with &ldquo;experimental&rdquo; values](notebooks/Ex_4_MSapp.ipynb)
-- [Example 5 - Using the pyscses with real data](notebooks/Ex_5_real_data.ipynb)
-- [Example 6 - Storing the output data](notebooks/Ex_6_store_data.ipynb)
+Each notebook can be viewed using [nbviewer](https://nbviewer.jupyter.org) by following each [nbviewer]() link.
+- [Theory](notebooks/Theory.ipynb) ([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Theory.ipynb))
+- [Setting up the notebook](notebooks/Setting_up.ipynb) ([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Setting_up.ipynb))
+- [Running the calculation](notebooks/Running.ipynb)([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Running.ipynb))
+- [Example 1 - continuum vs. site explicit and boundary conditions](notebooks/Ex_1_BC.ipynb)([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Ex_1_BC.ipynb))
+- [Example 2 - Mott-Schottky vs. Gouy-Chapman conditions](notebooks/Ex_2_MSGC.ipynb)([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Ex_2_MSCG.ipynb))
+- [Example 3 - Calculating the grain boundary resistivity and activation energy](notebooks/Ex_3_Res.ipynb)([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Ex_3_Res.ipynb))
+- [Example 4 - Comparison with &ldquo;experimental&rdquo; values](notebooks/Ex_4_MSapp.ipynb)([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Ex_4_MSapp.ipynb))
+- [Example 5 - Using the pyscses with real data](notebooks/Ex_5_real_data.ipynb)([nbviewer](https://nbviewer.jupyter.org/github/bjmorgan/pyscses/blob/master/userguides/notebooks/Ex_5_real_data.ipynb))
[…]
