### 2019-03-08 deddf23
Fixing spacing in userguide README
