### 2019-03-08 d58471a
Added userguide links to top level README
