### 2019-03-08 #1209 @bjmorgan
> The new contributing guidelines look good. I would suggest converting the "Issue Tracker" and "pull request" text into links.

Done https://github.com/bjmorgan/pyscses/commit/1fbab4696bd4b04ea26322891f843f071ee52bba
