### 2019-03-08 359ece4
first draft edits to paper.md
