+++ paper/paper.md
@@ -25,51 +25,36 @@ bibliography: paper.bib
 ---
 
 # Summary
+Solid electrolytes are often crystalline materials with high ionic conductivities, often used in solid state electrochemical devices. Most crystalline materials are characterised by defects, imperfections in the crystal that violate the perfect symmetries. Many solid electrolytes are polycrystalline with grain boundaries separating individual locally ordered regions with different orientations. In ionic crystals, thermodynamic arguments show that charged point defects will segregate to, or away from, interfaces and grain boundaries, and that there will be an associated redistribution of point defects in the regions adjacent. This redistribution leads to the regions having a net charge - known as space charge regions [@Maier_BerBunsenges1984; @Maier_JPhysChemSol1985; @Maier_SolStatIonics2003; @ChiangEtAl_ApplPhysLett1996; @KimAndMaier_JElectrochemSoc2002]. The segregation of defects can produce local defect concentrations that strongly deviate from the average &ldquo;bulk&rdquo; values in a material [@FleigAndMaier_SolidStateIonics1996], which can result in large variations in ionic conductivities[@JamnikAndMaier_SolidStateIonics1999]. ``pyscses`` is an open-source Python package for modelling ionic space-charge regions in crystalline solids. While the theory of space charge formation is well documented, there is currently no resources available for calculating space charge properties in solid electrolytes using explicitly defined defect positions and defect segregation energies. 
 
-``pyscses`` is an open-source Python package for modelling ionic space-charge regions in crystalline solids. Ionic space-charges are microscopic regions with a local excess or deficiency of charged point defects, giving these regions a net charge [@Frenkel_KinTheLiq1946]. Space-charge regions can form at crystallographic interfaces, such as grain boundaries or heterointerfaces between two different materials. The formation of space-charges is due to segregation of charged defects to, or from, these interfaces, and the associated redistribution of defects in adjacent regions of the crystal, due to defect&ndash;defect electrostatic interactions [@Maier_BerBunsenges1984; @Maier_JPhysChemSol1985; @Maier_SolStatIonics2003; @ChiangEtAl_ApplPhysLett1996; @KimAndMaier_JElectrochemSoc2002]. 
-
-The segregation of defects to, or from, space-charge regions can produce local defect concentrations that strongly deviate from the average &ldquo;bulk&rdquo; values in a material [@FleigAndMaier_SolidStateIonics1996]. This can have significant consequences for key material properties. For example, in solid electrolytes, where ionic defects act as mobile charge carriers, space-charge formation at grain boundaries is associated with large changes in ionic conductivity, both parallel and perpendicular to the grain boundaries, due to effects such as enhanced defect concentrations in the grain boundary core [@JamnikAndMaier_SolidStateIonics1999], or reduced defect concentrations in adjacent &ldquo;depletion&rdquo; space-charge regions. In the case of solid electrolytes, understanding space-charge formation at grain boundaries and interfaces is a key challenge in developing a theoretical description of the role of crystalline microstructure on macroscopic ion transport (ionic conductivities) [@KimEtAl_PhysChemChemPhys2003].
-
+# The numerical model
[…]
