### 2019-03-08 ce60e5e
first draft edits to paper.md
