+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Solid electrolytes are often crystalline materials with high ionic conductivities, often used in solid state electrochemical devices. Most crystalline materials are characterised by defects, imperfections in the crystal that violate the perfect symmetries. Many solid electrolytes are polycrystalline with grain boundaries separating individual locally ordered regions with different orientations. In ionic crystals, thermodynamic arguments show that charged point defects will segregate to, or away from, interfaces and grain boundaries, and that there will be an associated redistribution of point defects in the regions adjacent. This redistribution leads to the regions having a net charge - known as space charge regions [@Maier_BerBunsenges1984; @Maier_JPhysChemSol1985; @Maier_SolStatIonics2003; @ChiangEtAl_ApplPhysLett1996; @KimAndMaier_JElectrochemSoc2002]. The segregation of defects can produce local defect concentrations that strongly deviate from the average &ldquo;bulk&rdquo; values in a material [@FleigAndMaier_SolidStateIonics1996], which can result in large variations in ionic conductivities[@JamnikAndMaier_SolidStateIonics1999]. ``pyscses`` is an open-source Python package for modelling ionic space-charge regions in crystalline solids. While the theory of space charge formation is well documented, there is currently no resources available for calculating space charge properties in solid electrolytes using explicitly defined defect positions and defect segregation energies. 
+Solid electrolytes are often crystalline materials with high ionic conductivities, often used in solid state electrochemical devices. Most crystalline materials are characterised by defects, imperfections in the crystal that violate the perfect symmetries. Many solid electrolytes are polycrystalline with grain boundaries separating individual locally ordered regions with different orientations. In ionic crystals, thermodynamic arguments show that charged point defects will segregate to, or away from, interfaces and grain boundaries, and that there will be an associated redistribution of point defects in the regions adjacent. This redistribution leads to the regions having a net charge - known as space charge regions [@Maier_BerBunsenges1984; @Maier_JPhysChemSol1985; @Maier_SolStatIonics2003; @ChiangEtAl_ApplPhysLett1996; @KimAndMaier_JElectrochemSoc2002]. The segregation of defects can produce local defect concentrations that strongly deviate from the average &ldquo;bulk&rdquo; values in a material [@FleigAndMaier_SolidStateIonics1996], which can result in large variations in ionic conductivities[@JamnikAndMaier_SolidStateIonics1999]. It is therefore important to have a good understanding of defect behaviour and space charge formation to improve ionic conductivities in solid electrolytes for use in electrochemical devices. ``pyscses`` is an open-source Python package for modelling ionic space-charge regions in crystalline solids. While the theory of space charge formation is well documented, there is currently no resources available for calculating space charge properties in solid electrolytes using explicitly defined defect positions and defect segregation energies. 
 
 # The numerical model
 ``pyscses`` considers simple one-dimensional models of crystallographic interfaces, and calculates equilibrium defect distributions by solving a modified Poisson-Boltzmann equation [@Maier_ProgSolStatChem1995; @DeSouzaEtAl_SolidStateIonics2011]. The driving force for point-defect segregation to or from the interface is described by sets of defect segregation energies, $\Delta E_\mathrm{seg}^{i,x}$, defined as
