### 2019-03-13 #1209 @georgiewellock
To address the comments on the userguides:
> * [ ]  The userguide looks pretty good overall, you should fill out or get rid of the README file there though. Also, you should look into [nbsphinx](https://nbsphinx.readthedocs.io/en/0.4.2/), which will allow you to post the userguide as part of your RTD documentation. People shouldn't have to clone the repository to access that guide, since it's probably the more useful part of the documentation than the API. Let me know if you need examples of how this is done.

> Another alternative you may want to consider is mybinder. The service isn't perfect and does go down, so I would definitely only do that in addition to RTD/nbviewer, but it's really nice to be able to execute the notebooks directly online. That's just a suggestion though, not a hard request for this paper.

> To run the examples and tests I need certain notebooks that are in the repo that I
> can't download all together unless I clone the repo.
>
> Is there a better way to do this? you should put a note or description on how to get the examples and tests. Since we can't run the notebooks from github directly it would be useful to add a comment on where in the repo are located instead of putting "definitions and example code can be found in the `userguide`." where userguide is a specific link.

we have added nbviewer links to the top level README.md as discussed:

> Having said that, I do think nbviewer is also a good option if you dislike my suggestion. If you decide to stick with nbviewer, I would definitely add that to both your top-level README.md and your sphinx index.rst.
>

in this commit: https://github.com/bjmorgan/pyscses/commit/d58471a1a62a3e234c234bfc5ac310cd612cfbae

The userguides go through the theory of how to run the calculations, and if users want to run the calculations themselves I think it is useful to clone the whole repository.
