### 2019-03-13 #1209 @georgiewellock
I have run through the notebooks and fixed the errors that were flagged.

> * [ ]  Make sure the test notebooks actually run successfully; I ran into a couple of (easy-to-fix) issues such as needing to install pandas (that should just be documented at the top of the notebook), having to make the directory for storing outputs, etc.

> Make sure that all your notebooks run clean and it would be useful to have information on the time it takes to run some of the examples.
> - I open an issue on the repo so you can give me an estimate on the timing and maybe on the future add it to the documentation, since I've been running Example 1 for about 15 min and I don't know if this is the expected behavior.

Addressed here: https://github.com/bjmorgan/pyscses/commit/b845a5d21a2748867977e5c23f38f7c6b9f168fd
