### 2019-03-13 #1209 @georgiewellock
Going through the comments on this thread I believe we have addressed all of the software issues. We are currently redrafting the software paper which should be updated in the very near future and rewriting the summary in the README to clearly state what the problem is designed to solve.

- [ ] Update release version after review
- [ ] Updates to software paper
- [ ] Updates to top level README
