### 2019-03-17 6819dd1
Updated paper.md
