+++ paper/paper.bib
@@ -1,3 +1,95 @@
+
+@article{SandveEtAl_PLoSComputBiol2013,
+	Author = {Sandve, Geir Kjetil and Nekrutenko, Anton and Taylor, James and Hovig, Eivind},
+	Date-Added = {2017-06-03 16:49:37 +0000},
+	Date-Modified = {2017-06-05 13:38:50 +0000},
+	Journal = {PLoS Comput. Biol.},
+	Month = oct,
+	Number = {10},
+	Pages = {e1003285--4},
+	Title = {Ten Simple Rules for Reproducible Computational Research},
+	Volume = {9},
+	Year = {2013},
+        DOI = {10.1371/journal.pcbi.1003285}}
+
+@article{LindmanEtAl_SolStatIonics2013a,
+	Author = {Lindman, Anders and Helgee, Edit E. and Wahnstr\"{o}m, G\"{o}ran},
+	Date-Added = {2019-03-17 10:35:54 +0000},
+	Date-Modified = {2019-03-17 10:36:28 +0000},
+	Journal = {Sol. Stat. Ionics},
+	Month = may,
+	Pages = {1--5},
+	Title = {Theoretical modeling of defect segregation and space-charge formation in the {B}a{Z}r{O}$_{3}$ $(210)[001]$ tilt grain boundary},
+	Year = {2013},
+        DOI = {10.1016/j.ssi.2013.04.008}}
+
+
+@article{NymanEtAl_ApplPhysLett2012,
+	Author = {Nyman, Joakim B. and Helgee, Edit E. and Wahnstr\"{o}m, G\"{o}ran},
+	Date-Added = {2019-03-17 10:34:25 +0000},
+	Date-Modified = {2019-03-17 10:34:51 +0000},
+	Journal = {Appl. Phys. Lett.},
+	Month = feb,
+	Number = {6},
+	Pages = {061903--4},
+	Title = {Oxygen vacancy segregation and space-charge effects in grain boundaries of dry and hydrated {B}a{Z}r{O}$_3$},
+	Volume = {100},
+	Year = {2012},
+        DOI = {10.1063/1.3681169}}
+
+@article{GuoAndWaser_ProgMaterSci2006,
+	Author = {Guo, X. and Waser, R.},
+	Date-Added = {2019-03-10 08:32:13 +0000},
+	Date-Modified = {2019-03-10 08:32:13 +0000},
+	Journal = {Progress in Materials Science},
+	Month = feb,
+	Number = {2},
+	Pages = {151--210},
+	Title = {Electrical properties of the grain boundaries of oxygen ion conductors: Acceptor-doped zirconia and ceria},
+	Volume = {51},
+	Year = {2006},
+        DOI = {10.1016/j.pmatsci.2005.07.001}}
+
+
+@article{Franceschetti_SolStatIonics1981,
+	Author = {Franceschetti, D. R.},
+	Date-Added = {2019-03-17 10:33:01 +0000},
+	Date-Modified = {2019-03-17 10:33:01 +0000},
+	Journal = {Sol. Stat. Ionics},
+	Number = {1},
+	Pages = {39--42},
+	Title = {Local Thermodynamic Formalism for Space-Charge in Ionic-Crystals},
+	Volume = {2},
+	Year = {1981},
+        DOI = {10.1016/0167-2738(81)90017-5}}
+
+
+@article{MahatoEtAl_ProgMaterSci2015,
+	Author = {Mahato, Neelima and Banerjee, Amitava and Gupta, Alka and Omar, Shobit and Balani, Kantesh},
+	Date-Added = {2019-03-15 13:58:11 +0000},
+	Date-Modified = {2019-03-15 13:58:11 +0000},
+	Journal = {Progress in Materials Science},
+	Month = jul,
+	Number = {C},
[…]
