+++ paper/paper.md
@@ -25,29 +25,33 @@ bibliography: paper.bib
 ---
 
 # Summary
-Solid electrolytes are often crystalline materials with high ionic conductivities, often used in solid state electrochemical devices. Most crystalline materials are characterised by defects, imperfections in the crystal that violate the perfect symmetries. Many solid electrolytes are polycrystalline with grain boundaries separating individual locally ordered regions with different orientations. In ionic crystals, thermodynamic arguments show that charged point defects will segregate to, or away from, interfaces and grain boundaries, and that there will be an associated redistribution of point defects in the regions adjacent. This redistribution leads to the regions having a net charge - known as space charge regions [@Maier_BerBunsenges1984; @Maier_JPhysChemSol1985; @Maier_SolStatIonics2003; @ChiangEtAl_ApplPhysLett1996; @KimAndMaier_JElectrochemSoc2002]. The segregation of defects can produce local defect concentrations that strongly deviate from the average &ldquo;bulk&rdquo; values in a material [@FleigAndMaier_SolidStateIonics1996], which can result in large variations in ionic conductivities[@JamnikAndMaier_SolidStateIonics1999]. It is therefore important to have a good understanding of defect behaviour and space charge formation to improve ionic conductivities in solid electrolytes for use in electrochemical devices. ``pyscses`` is an open-source Python package for modelling ionic space-charge regions in crystalline solids. While the theory of space charge formation is well documented, there is currently no resources available for calculating space charge properties in solid electrolytes using explicitly defined defect positions and defect segregation energies. 
+`pyscses` is a Python package for modelling ionic space-charges in solid electrolytes. Its primary use is to calculate equilibrium distributions of point-charge atomic defects within one-dimensional &ldquo;Poisson-Boltzmann&rdquo;-like mean-field models. These calculations take as inputs a set of defect site positions, within a specific crystal structure, and the associated defect segregation energies. `pyscses` can also be used to calculate ionic transport properties (space-charge resistivities and activation energies) for these equilibrium defect distributions.
 
-# The numerical model
-``pyscses`` considers simple one-dimensional models of crystallographic interfaces, and calculates equilibrium defect distributions by solving a modified Poisson-Boltzmann equation [@Maier_ProgSolStatChem1995; @DeSouzaEtAl_SolidStateIonics2011]. The driving force for point-defect segregation to or from the interface is described by sets of defect segregation energies, $\Delta E_\mathrm{seg}^{i,x}$, defined as
-$$
-\Delta E_\mathrm{seg}^{i,x} = E_\mathrm{f}^{i,x} - E_\mathrm{f}^{i, \infty}
-$$
-i.e. the difference in defect formation energy for defect species $i$ at site $x$ compared to a reference site in the &ldquo;bulk&rdquo; of the crystal.
+# Scientific Context
+Crystalline solids consist of periodic arrangements of atoms that vibrate about fixed lattice sites. In most solids, atoms rarely move between lattice sites, and long ranged atomic diffusion is slow. Solid electrolytes are notable because they contain atoms that can rapidly move between lattice sites, giving high diffusion coefficients. This unusual property makes them useful in solid-state electrochemical devices such as fuel cells and solid-state batteries [@MahatoEtAl_ProgMaterSci2015; @ZhengEtAl_JPowerSources2018]. Many practical solid electrolytes are polycrystalline: they contain multiple crystalline domains, with varied orientations, that meet at grain boundaries. 
 
[…]
