### 2019-03-17 #1209 @bjmorgan
We have revised and restructured the `paper.md` https://github.com/bjmorgan/pyscses/commit/6819dd1a156cc0c70b19831416f779b3a12d2aa7

Overview of changes:
- We now have a short **Summary** section, that explains what `pyscses` is written to do, so that readers can check this without having to wade through the scientific background.
- We have moved the scientific context to a **Scientific Context** section. We feel that including this is helpful for readers who might find the software useful, but who are not clear exactly how this fits with their own research. We have rewritten this content to (hopefully) make it more accessible to readers with general materials science backgrounds, but who might not be familiar with the specific physics of this problem set.
- We have simplified the **Numerical Model** section. We feel that keeping some details about the exact model handled by `pyscses` is helpful for our target users (people interested in modelling ionic space-charges) who should be able to tell from scanning this paper whether `pyscses` is helpful for solving their models, without having to go to the repository and find the online userguide. We now give brief lists of models, calculated properties, and key approximations and limitations.
