### 2019-03-17 e2a315a
Reordered contributions / citing in README
