### 2019-03-17 #1209 @bjmorgan
We have revised the top level `README`: To avoid clogging up the practical instructions of how to install and then learn about *using* `pyscses` we have directed readers interested in an overview of the scientific context to the JOSS paper. https://github.com/bjmorgan/pyscses/commit/e2a315a96d05b156242cf1ecb51a6d56d06bbd21
