### 2019-03-17 #1209 @bjmorgan
Re: statement of need. Although the 1D Poisson-Boltzmann model is a fairly common numerical problem, we are not aware of any open source codes that solve this specifically for atomic defects in crystalline solids, or of any papers in this field that contain reproducible computational details. The updated paper contains a couple of sentences on this point.
