### 2019-03-18 #1209 @ncclementi
@bjmorgan and @georgiewellock

### Regarding the `paper.md`
I think you did a great job. My only suggestion would be to add in the summary a sentence related to the Statement of need. It is really well explained and clear by the end of the **scientific context** section, but I think if someone just ready the summary will miss this. Similarly I would add this to the README too, a short sentence that shows why you are different than the rest and why your software is a step forward in terms of reproducible research.

### Regarding the `tests`:
1. I wasn't able to run the notebooks, I got an error since the folder `generated_outputs` doesn't exist. You should create this directory before the `numpy.savetxt()` line.
2. In the notebooks I'd add a header line that tells me what is this test about and how long it'll take to run. Similar to what you did in the examples.
3. I have no problems running the unit tests but it would be useful having some documentation on them that tells me what are we testing.
