### 2019-03-19 #1209 @vyasr
I concur with the comments that @ncclementi made, those are largely identical to the issues I would have raised. One minor comment, I'd remove one of the logo files (you currently have two in your repo (one at the root and one in the resources folder). Aside from that, I think everything looks good!
