### 2019-03-19 f60eda3
Moved project logo to resources
