### 2019-03-19 eabedb0
Revert "Reordered contributions / citing in README"
This reverts commit e2a315a96d05b156242cf1ecb51a6d56d06bbd21.
