### 2019-03-19 19c55eb
Fixing deleted logo.png
