### 2019-03-19 bjmorgan/pyscses#11 @bjmorgan
JOSS paper updated in https://github.com/bjmorgan/pyscses/commit/6819dd1a156cc0c70b19831416f779b3a12d2aa7.
