### 2019-03-19 #1209 @bjmorgan
@vyasr These logo files are different. The `logo.png` in the root directory is for display in the README as a webpage (e.g. on GitHub). The `pyscses logo.pdf` in the `resources` folder is the original vector image for the logo, saved for any future use.
