### 2019-03-19 37b1fc6
Edited citation link syntax
