### 2019-03-19 dca7cb8
Fix for citation links
