### 2019-03-19 #1209 @bjmorgan
@ncclementi

> My only suggestion would be to add in the summary a sentence related to the Statement of need. It is really well explained and clear by the end of the scientific context section, but I think if someone just ready the summary will miss this. Similarly I would add this to the README too, a short sentence that shows why you are different than the rest and why your software is a step forward in terms of reproducible research.

We have added the paragraph describing the Statement of Need to the top level README. https://github.com/bjmorgan/pyscses/commit/4255c7df2a71b2b789e363799b56446c9e20fa9a

We have rearranged the `paper.md` so that the Statement of Need is now in the Summary, instead of the Scientific Context section. https://github.com/bjmorgan/pyscses/commit/05d2084e67668c9724d1b2465ac6c1967b8a663f
