+++ tests/test_notebooks/README.md
@@ -1,0 +1,20 @@
+# Test Notebooks
+
+TODO
+
+Running these notebooks requires additional Python dependencies. These can be installed using
+```
+cd tests/test_notebooks
+pip install -r requirements.txt
+```
+
+## Notebooks
+
+TODO
+
+## Running times
+
+Some of these notebooks contain relatively intensive calculations. If you rerun these from scratch, in some cases please be prepared to wait for the calculation to reach completion.  
+Typical running times for each notebook (4 GHz i7 iMac):
+TODO
+
