+++ requirements.txt
@@ -1,4 +1,3 @@
-pandas
 numpy
 mpmath
 scipy
