+++ userguides/requirements.txt
@@ -1,4 +1,3 @@
+-r ../requirements.txt
 pandas
-numpy
-mpmath
-scipy
+jupyter
