### 2019-03-19 74af68b
Added check for missing generated_outputs directories
