+++ userguides/requirements.txt
@@ -1,3 +1,4 @@
 -r ../requirements.txt
 pandas
 jupyter
+matplotlib
