### 2019-03-19 #1209 @vyasr
Got it.

@labarba I'm happy with the package now. Let me know if I should copy-edit the paper, otherwise I think I'm all set on my end.
