### 2019-03-19 6590f8c
README and test notebooks updated with a description and an approximate run time.
