### 2019-03-19 #1209 @ncclementi
On my end I'd like the test notebooks to run, as I mention above there are few little problems.
Regarding the `paper.md` it looks great now!
