### 2019-03-19 bd7f5d0
nbviewer links added
