### 2019-03-19 #1209 @georgiewellock
> ### Regarding the `tests`:
> 1. I wasn't able to run the notebooks, I got an error since the folder `generated_outputs` doesn't exist. You should create this directory before the `numpy.savetxt()` line.

A check has been added for the `generated_outputs` directory and creates it if it doesn't exist in commit: https://github.com/bjmorgan/pyscses/commit/74af68b52c91ba9e955ef918f8b6081bb86acec4

> 2. In the notebooks I'd add a header line that tells me what is this test about and how long it'll take to run. Similar to what you did in the examples.

Descriptions and approximate run times have been added to each test notebook and the associated README in commit: https://github.com/bjmorgan/pyscses/commit/6590f8cbd0b72866a67901b0211d1565207cde0f
