### 2019-03-19 #1209 @ncclementi
Added a comment on the paragraph regarding the tests
https://github.com/bjmorgan/pyscses/commit/6590f8cbd0b72866a67901b0211d1565207cde0f#r32824525
