### 2019-03-19 9e1865d
rephrased to remove the repetitive calculation
