### 2019-03-19 #1209 @georgiewellock
I have rephrased the paragraph in commit https://github.com/bjmorgan/pyscses/commit/9e1865db501d23eb946651fb5545c751ae6ff204
