### 2019-03-19 #1209 @ncclementi
Test notebooks now run properly. I have nothing else to add or request on my end.
Great job @bjmorgan and @georgiewellock
