### 2019-03-19 #1209 @labarba
Editorial fixes:

¶2
typo: obtined >> obtained

page 2: comma after "i.e."

"a range of numerical models" ... did you mean what follows to be a bulleted list? If not, delete the hyphens and numerate in-sentence. If yes, work on the formatting. (But I think it will look better as a bulleted list.) Same for next paragraph.
