### 2019-03-19 580f4ae
Fixes for paper.md
cf. https://github.com/openjournals/joss-reviews/issues/1209#issuecomment-474549427
