### 2019-03-20 #1209 @labarba
Great. It's now time to make a tagged release (report release number here) and make an archive of the software in Zenodo (or similar service). Make sure to manually edit the Zenodo metadata so the title and author list matches the JOSS paper, then report the archive DOI here.
