### 2019-03-20 1c3911e
Bump to version 1.0.0
