+++ pyscses/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.9.3.3'
+__version__ = '1.0.0'
