### 2019-03-20 #1209 @bjmorgan
Version number bumped to [1.0.0](https://github.com/bjmorgan/pyscses/releases/tag/1.0.0) and released (also updated on [PyPI](https://pypi.org/project/pyscses/)).
The Zenodo doi for this version is [10.5281/zenodo.2599955](https://zenodo.org/account/settings/github/repository/bjmorgan/pyscses).
