### 2019-03-20 6f6ab97
Updated citation metadata in README
