### 2019-03-20 #1209 @labarba
Congratulations, @bjmorgan and @georgiewellock — your JOSS paper is published!

With huge thanks to our reviewers: @ncclementi, @vyasr — we couldn't do this without you 🙏
